style(windows): silence remaining MSVC C4305/C4244 warnings (follow-up) (#15)
Continuation of #14. Mechanical pass over the ~161 warning sites that
were skipped in the first batch. Improved the script to:

  - Handle assignment warnings whose column points into the RHS
    expression (scan backward for `=`).
  - Recognize template-dependent source types (`_Ty` from
    `std::vector<double>::operator[]`).

Also added a tiny targeted pass that suffixes `f` to double literals
inside `_PS_CONST(...)` SSE constant-table macros in
`inc/fast-GB/SSEApproxMath.h`.

Files touched (sites fixed):
  - src/f2dock/Docking.cpp                 48
  - inc/fast-GB/SSEApproxMath.h            21 (_PS_CONST literals)
  - src/PG-range/PG.cc                     15
  - src/fast-hydro/pseudoGsol.cpp          15
  - src/fast-PQ/PairingHeap.cpp             9
  - src/GB-rerank/pairingHeap.cpp           9
  - src/f2dock/F2Dock.cpp                   7
  - src/GB-rerank/GB-rerank.cpp             6
  - src/fast-resCont/resContFilter.cpp      4
  - inc/PG-range/PG.h                       1

~31 warning sites remain (complex multi-line expressions, std::printf
format-argument conversions, and a few corner cases). They are deferred
to a future manual pass.

libcvc warnings (`build/_deps/libcvc-src/...`) remain untouched and will
be addressed upstream.

+++ inc/PG-range/PG.h
@@ -54,7 +54,7 @@ struct Point {
     dx = a.x - b.x;
     dy = a.y - b.y;
     dz = a.z - b.z;
-    res = dx * dx + dy * dy + dz * dz;
+    res = (float)(dx * dx + dy * dy + dz * dz);
     return res;
   }
 };
